refactor(global): move global api to inside of error package and effected class
