+++ src/main/java/online/lifeasgame/global/error/api/model/ViolationItem.java
@@ -1,3 +1,3 @@
-package online.lifeasgame.global.api.model;
+package online.lifeasgame.global.error.api.model;
 
 public record ViolationItem(String property, String msg, String invalid) {}
